Adds functional animations to sprites.
Sprites now have a playAnimation function that can play animations that are loaded into it's animation set.
Each sprite now contains a cpAnimSet pointer and a cpTexture pointer. A Sprite has two constructor's, one with
a cpAnimSet pointer, and one with a cpTexture pointer. If you use the constructor with just a Texture, then it
is instantiated as a non animated object. Each update and render call doesn't need to deduce the current frame.
If it is instantiated with a cpAnimSet pointer then all the functionality of the animation members are put in use
per frame.