Rips out the whole PhysSprite/AObj architecture.
I also got rid of the god-awful macro bug. NEVER put function calls
directly in a macro that is removable depending on the build configuration.
I spent literally all day trying to find this one bug, and it was because assertions
are not enabled in the final and release builds. There was one line like
|TALGA_ASSERT(glfwInit(), "glfw failed to init... blblablablf" and it was removing
the glfw initialization completely.